refactor(gen): Moved gen folder inside core and removed snackbar implementation in auth-servicee
