+++ Repository/authentication/lib/src/core/widgets/button/back_button.dart
@@ -3,7 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../gen/assets.gen.dart';
+import '../../gen/assets.gen.dart';
+
 
 class AppBackButton extends StatelessWidget {
   const AppBackButton({
