+++ Repository/authentication/lib/src/features/authentication/change_password/presentation/pages/change_password_page.dart
@@ -12,7 +12,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../../gen/assets.gen.dart';
+import '../../../../../core/gen/assets.gen.dart';
 import '../../../../../core/notifiers/text_edtitting_controller_listener.dart';
 import '../../../../../core/services/routes/route_generator.dart';
