+++ Repository/authentication/lib/src/features/authentication/root/presentation/widgets/build_back_button.dart
@@ -3,7 +3,8 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 
-import '../../../../../../gen/assets.gen.dart';
+import '../../../../../core/gen/assets.gen.dart';
+
 
 class BuildBackButton extends StatelessWidget {
   const BuildBackButton({super.key});
