+++ Repository/authentication/lib/src/features/authentication/sign_up/presentation/pages/sign_up_success_page.dart
@@ -5,7 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../../../../gen/assets.gen.dart';
+import '../../../../../core/gen/assets.gen.dart';
+
 
 class SuccessPage extends StatefulWidget {
   const SuccessPage({super.key});
