+++ Repository/authentication/pubspec.yaml
@@ -56,7 +56,12 @@ dev_dependencies:
   very_good_analysis: ^5.1.0
 
 
+
+flutter_gen:
+  output: lib/src/core/gen
+
 flutter:
+
 
   uses-material-design: true
